feat(bills): render bills and tipcards over all content at the app root
In the v2 tab-bar UI, cash bills and tipcards were rendered inside
ScanScreen's ZStack. A bill set by a push or deep link (e.g. an incoming
cash link) while the user was on the Wallet/Chat/Tip Card tab wrote to
session.billState but rendered in the unmounted Scan tab, so nothing
appeared on screen.

Hoist the bill/tipcard surface — the BillCanvas, the primary/secondary
bill actions, the bill designer, and the received-cash / send-tip sheets
— into a new app-root BillOverlayView, mounted once by ContainerScreen
above both the v1 scanner and the v2 HomeTabView. Driven entirely by the
app-scoped session.billState, a bill now renders over whatever surface is
showing, mirroring Android's app-root BillOverlay. ScanScreen keeps its
showControls gate so its scanner chrome still hides while a bill is up.

Details:
- The overlay keeps ignoresSafeArea() on its inner surface, not its root:
  as a sibling of the tab NavigationStacks, an overlay whose root ignored
  the safe area flattened the iOS 26 Liquid Glass of the nav bars beneath
  it. Nested one level down, the bill fills the screen and the bars keep
  their glass.
- A scrim is drawn behind the bill when it is presented over a non-scanner
  surface (a wallet/chat push, a received cash link), ramping in with the
  bill's slide. Over the scanner the live camera stays visible instead —
  gated by the new Session.isScannerForeground, set by ScanScreen.
- The floating tab bar hides whenever the active tab has pushed a screen
  (e.g. Wallet -> Currency Info -> Give) or while a bill is presenting, so
  the pushed screen / bill owns the full height instead of the bar
  hovering over it.
